Versione 0.0.750
Aggiunta la libreria ORMDroid per la gestione del DB locale. Decisione presa dopo aver constatato la difficile gestione del DB usando la struttura nativa offerta da Android, più che altro perche', facendo alcune prove con diversi utenti, mi sono ritrovato un funzionamento anomalo.

+++ Interventix/src/com/federicocolantoni/projects/interventix/task/GetUtenteAsyncTack.java
@@ -43,7 +43,7 @@ public class GetUtenteAsyncTack extends AsyncTask<Long, Void, Utente> {
 			utente.setCognome(cursor.getString(cursor.getColumnIndex(UtenteDB.Fields.COGNOME)));
 			utente.setConflitto(cursor.getInt(cursor.getColumnIndex(UtenteDB.Fields.CONFLITTO)) == 1 ? true : false);
 			utente.setEmail(cursor.getString(cursor.getColumnIndex(UtenteDB.Fields.EMAIL)));
-			utente.setIdUtente(cursor.getLong(cursor.getColumnIndex(UtenteDB.Fields.ID_UTENTE)));
+			utente.setIdUtente(cursor.getInt(cursor.getColumnIndex(UtenteDB.Fields.ID_UTENTE)));
 			utente.setNome(cursor.getString(cursor.getColumnIndex(UtenteDB.Fields.NOME)));
 			utente.setPassword(cursor.getString(cursor.getColumnIndex(UtenteDB.Fields.PASSWORD)));
 			utente.setRevisione(cursor.getLong(cursor.getColumnIndex(UtenteDB.Fields.REVISIONE)));
